+++ src/components/main-container/main-container.tsx
@@ -10,6 +10,7 @@ const SeasonPage = lazy(() => import("../season/season-page"));
 const SeriesPage = lazy(() => import("../series/series-page"));
 const ShopPage = lazy(() => import("../shop-guide/shop-page"));
 const TracksPage = lazy(() => import("../content/tracks-page"));
+const HistoryPage = lazy(() => import("../history/history-page"));
 
 function MainContainer() {
   usePageTracking();
@@ -35,6 +36,7 @@ function MainContainer() {
           <Route path={ETabs.MyCars} component={CarsPage} />
           <Route path={ETabs.MyTracks} component={TracksPage} />
           <Route path={ETabs.ShopGuide} component={ShopPage} />
+          <Route path={ETabs.History} component={HistoryPage} />
 
           <Route>
             <Redirect to={ETabs.MySeason} />