+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the **"my-racing-planner"** project will be documented in
 
 ## 0.7.0
 
+- Added Tracks Usage history page
 - Removed Info Link to iRacing members site on Content tables
 
 ## 0.6.4